"ssh-keygen -A -m PEM" does not work.
https://github.com/random-archer/mkinitcpio-systemd-tool/issues/83

The manpage of ssh-keygen describes option "-A" as
"For each of the key types (rsa, dsa, ecdsa and ed25519) for which host keys do not exist, generate the host keys [...]."

But since the host keys in /etc/ssh/ already exists (because they are generated on the first invocation of sshd), nothing will be generated/regenerated/converted. Id2ndR suggested the for-loop method instead.